refactoring (moved function to db_functions file)
