+++ dashboard/db_functions.py
@@ -37,3 +37,20 @@ def get_topic_dict() -> dict:
             cur.execute(query)
             res = cur.fetchall()
     return {topic['topic_name']: topic['topic_id'] for topic in res}
+
+def get_scores_topic(topic_name: str) -> dict:
+    """Returns a dictionary containing the polarity scores for a given topic """
+
+    topic_name = topic_name.strip().title()
+    with create_connection() as conn:
+        select_data = """SELECT t.topic_name, s.source_name, a.content_polarity_score, a.title_polarity_score, a.date_published FROM article a
+        INNER JOIN article_topic_assignment ata ON a.article_id = ata.article_id 
+        INNER JOIN topic t ON ata.topic_id = t.topic_id 
+        INNER JOIN source s ON a.source_id = s.source_id
+        WHERE t.topic_name = %s 
+        """
+        with conn.cursor() as curr:
+            curr.execute(select_data,(topic_name, ))
+            res = curr.fetchall()
+
+    return res
